requirements.txt
# PDF Processing
PyMuPDF>=1.24.0

# gRPC
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.1

# HTTP Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
httpx>=0.25.0
python-multipart>=0.0.6

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Logging
structlog==23.2.0
